Convert readonly trigger related code to condition monitor
Cut and paste methods related to monitoring/trigger a task from ReadOnly file to
ConditionMonitor file
